Fix float-to-bool conversion in Pocket
The `Pocket` class includes a facility for automatic conversion of data
types upon retrieval. While this functionality is not a good design in
the first place, it should at least work consistently.

With the current code, a small-but-nonzero floating point number stored
as string will convert to `False`. This is dangerously inconsistent with
Python's own type conversion rules.

Part of the reason for this is a broken regex for floating point numbers
stored as string. Drop this regex, in favor of "try and ask for
forgiveness" style conversion: simply call `float()` on the string, and
catch any `TypeError` or `ValueError` that comes out of it.

Also drop the mistaken logic of first rounding the floatign point number
to an int before comparing to zero.

Includes updated and extended unit tests to cover this functionality.