NonVolatilePocket: Add tests that exercise timing
One of `NonVolatilePocket`'s main functionalities is "rate limiting" of
writes to disk. This involves a change monitoring thread that is
supposed to fire only once every X seconds, whenever changes are written
to the `Pocket`.

This commit adds a number of tests that exercise this behavior, in
combination with the other APIs like `erase()`:

- Before 407a9dc, the `NonVolatilePocket` change monitoring thread
contained a race condition that would trigger in combination with
`erase()`. We specifically cover this edge cas in `test_erase_timing`.

- The `erase()` function provides an argument to kill the change
monitoring thread after erasure. Test this, in
`test_erase_no_restart()`.

- Test the rate limiting behavior in `test_save_rate_limit()`.

- Subpockets should propagate change signals to their parents. Test this
implementation in `test_save_subpocket()`.

- Pockets should still safeguard their data on program exit or
destruction. Test both scenarios, in  `test_save_upon_*()`. Note that
this actually exposes a bug: destroying a `NonVolatilePocket` does *not*
write the data to disk.